Time: 154 ms (67.39%), Space: 111.2 MB (13.24%) - LeetHub

+++ 3075-maximize-happiness-of-selected-children/3075-maximize-happiness-of-selected-children.cpp
@@ -29,7 +29,7 @@ public:
             newHappiness[i] = happiness[i] - i;
         }
 
-        int maxHappiness = 0, i = 0;
+        long long maxHappiness = 0, i = 0;
         while(k--){
             maxHappiness = max(maxHappiness, maxHappiness + newHappiness[i++]);
         }
